move functions in same file
